CORTX-31129 Moved BSIZE_LAYOUT_MAP to Layout.py.
