+++ libs/motr/layouts.py
@@ -32,4 +32,8 @@ bsize_list = ['4K', '8K', '16K', '32K', '64K', '128K', '256K', '512K',
 layout_ids = ['1', '2', '3', '4', '5', '6', '7', '8',
               '9', '10', '11', '12', '13', '14']
 
+BSIZE_LAYOUT_MAP = {'4K': 1, '8K': 2, '16K': 3, '32K': 4, '64K': 5, '128K': 6, '256K': 7,
+                    '512K': 8, '1M': 9, '2M': 10, '4M': 11, '8M': 12, '16M': 13, '32M': 14}
 
+LAYOUT_BSIZE_MAP = {1: '4K', 2: '8K', 3: '16K', 4: '32K', 5: '64K', 6: '128K', 7: '256K',
+                    8: '512K', 9: '1M', 10: '2M', 11: '4M', 12: '8M', 13: '16M', 14: '32M'}
